requirements.txt
bs4
requests
pytest
flask
openai
chromadb
python-dotenv
spacy
vaderSentiment
boto3
newspaper3k
lxml_html_clean